service/wip -> added files middleware, custom logging; refactor errors
